# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Prime Link OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [users]
      summary: Update user online status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: assignedTo
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        online:
          type: boolean
        lastActive:
          type: ["string", "null"]
        performanceScore:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - status
        - online
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
      required:
        - name
        - email
        - password
        - role
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        status:
          type: string
        performanceScore:
          type: ["number", "null"]
    UpdateStatusBody:
      type: object
      properties:
        online:
          type: boolean
      required:
        - online
...
    Task:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: ["integer", "null"]
        assignedTo:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
        status:
          type: string
        deadline:
          type: ["string", "null"]
        progress:
          type: ["integer", "null"]
        resultUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - priority
        - status
        - createdAt
    CreateTaskBody:
      type: object
      properties:
        clientId:
          type: ["integer", "null"]
        assignedTo:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        deadline:
          type: ["string", "null"]
      required:
        - title
        - priority
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        assignedTo:
          type: ["integer", "null"]
        deadline:
          type: ["string", "null"]
        progress:
          type: ["integer", "null"]
        resultUrl:
          type: string
...